server/requirements.txt
Flask==3.0.3
ipykernel==6.29.5
matplotlib==3.9.1
numpy==2.0.0
openai==1.35.10
pandas==2.2.2
python-dotenv==0.19.0
requests==2.32.3
urllib3==2.2.2
xmltodict==0.13.0
PyJWT==2.8.0
pydantic==2.8.2
Flask-PyMongo
bcrypt